requirements.txt
flask==2.3.3
werkzeug==2.3.4
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.0
pillow==10.0.0
